python/requirements.txt
# Python 3.13+ required
# Install: pip install -r requirements.txt

# Gmail Watcher (silver tier)
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0

# File system watcher (bronze tier — watchdog optional, polling used by default)
watchdog>=4.0.0
